moved worker logic in dedicated module
